rename XROffset -> BioIkObjectiveWithOffset
